backend/backend-delopy1.016/requirements.txt
bcrypt==4.1.2
boto3==1.34.0
botocore==1.34.0
fastapi==0.104.1
gunicorn==21.2.0
numpy==1.25.0
pandas==2.1.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
pymongo==4.6.0
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.0
sqlalchemy==2.0.23
uvicorn[standard]==0.24.0
